compute/pyproject.toml
[project]
name = "rosetta-compute"
version = "0.1.0"
description = "Modular CDC engine for data streaming"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jpype1>=1.6.0",
    'pydbzengine @git+https://github.com/memiiso/pydbzengine.git',
    "psycopg2-binary>=2.9.9",
    "duckdb>=1.4.4",
    "python-dotenv>=1.0.0",
    "cryptography==42.0.0",
    "httpx==0.28.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pyjwt>=2.11.0",
    "pandas>=3.0.0",
    "pyarrow>=15.0.0",
    "redis>=5.0.0",
    "pytz>=2025.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
rosetta-compute = "python -m compute.main"

[tool.uv]
native-tls = true
required-environments = [
    "sys_platform == 'linux'",
    "sys_platform == 'darwin'",
]
